Create new pkg/types and move Any there so that Any can be used across parts of the project
