+++ pkg/plugin/group/integration_test.go
@@ -9,10 +9,11 @@ import (
 	"time"
 
 	plugin_base "github.com/docker/infrakit/pkg/plugin"
-	"github.com/docker/infrakit/pkg/plugin/group/types"
+	group_types "github.com/docker/infrakit/pkg/plugin/group/types"
 	"github.com/docker/infrakit/pkg/spi/flavor"
 	"github.com/docker/infrakit/pkg/spi/group"
 	"github.com/docker/infrakit/pkg/spi/instance"
+	"github.com/docker/infrakit/pkg/types"
 	"github.com/stretchr/testify/require"
 )
 
@@ -39,8 +40,8 @@ func flavorPluginLookup(_ plugin_base.Name) (flavor.Plugin, error) {
 	return &testFlavor{}, nil
 }
 
-func minionProperties(instances int, instanceData string, flavorInit string) *plugin_base.Any {
-	return plugin_base.AnyString(fmt.Sprintf(`{
+func minionProperties(instances int, instanceData string, flavorInit string) *types.Any {
+	return types.AnyString(fmt.Sprintf(`{
 	  "Allocation": {
 	    "Size": %d
 	  },
@@ -60,13 +61,13 @@ func minionProperties(instances int, instanceData string, flavorInit string) *pl
 	}`, instances, instanceData, flavorInit))
 }
 
-func leaderProperties(logicalIDs []instance.LogicalID, data string) *plugin_base.Any {
+func leaderProperties(logicalIDs []instance.LogicalID, data string) *types.Any {
 	idsValue, err := json.Marshal(logicalIDs)
 	if err != nil {
 		panic(err)
 	}
 
-	return plugin_base.AnyString(fmt.Sprintf(`{
+	return types.AnyString(fmt.Sprintf(`{
 	  "Allocation": {
 	    "LogicalIDs": %s
 	  },
@@ -110,7 +111,7 @@ func memberTags(id group.ID) map[string]string {
 
 func provisionTags(config group.Spec) map[string]string {
 	tags := memberTags(config.ID)
-	tags[configTag] = types.MustParse(types.ParseProperties(config)).InstanceHash()
+	tags[configTag] = group_types.MustParse(group_types.ParseProperties(config)).InstanceHash()
 
 	return tags
 }
@@ -435,7 +436,7 @@ func TestInstanceAndFlavorChange(t *testing.T) {
 		require.Equal(t, "updated init", inst.Init)
 
 		properties := map[string]string{}
-		err = plugin_base.AnyBytes([]byte(*inst.Properties)).Decode(&properties)
+		err = types.AnyBytes([]byte(*inst.Properties)).Decode(&properties)
 		require.NoError(t, err)
 		require.Equal(t, "data2", properties["OpaqueValue"])
 	}
