+++ pkg/plugin/group/types/types.go
@@ -9,6 +9,7 @@ import (
 	"github.com/docker/infrakit/pkg/plugin"
 	"github.com/docker/infrakit/pkg/spi/group"
 	"github.com/docker/infrakit/pkg/spi/instance"
+	"github.com/docker/infrakit/pkg/types"
 )
 
 // Spec is the configuration schema for the plugin, provided in group.Spec.Properties
@@ -27,13 +28,13 @@ type AllocationMethod struct {
 // InstancePlugin is the structure that describes an instance plugin.
 type InstancePlugin struct {
 	Plugin     plugin.Name
-	Properties *plugin.Any // this will be the Spec of the plugin
+	Properties *types.Any // this will be the Spec of the plugin
 }
 
 // FlavorPlugin describes the flavor configuration
 type FlavorPlugin struct {
 	Plugin     plugin.Name
-	Properties *plugin.Any // this will be the Spec of the plugin
+	Properties *types.Any // this will be the Spec of the plugin
 }
 
 // ParseProperties parses the group plugin properties JSON document in a group configuration.
@@ -50,7 +51,7 @@ func ParseProperties(config group.Spec) (Spec, error) {
 // UnparseProperties composes group.spec from id and props
 func UnparseProperties(id string, props Spec) (group.Spec, error) {
 	unparsed := group.Spec{ID: group.ID(id)}
-	any, err := plugin.AnyValue(props)
+	any, err := types.AnyValue(props)
 	if err != nil {
 		return unparsed, err
 	}
@@ -104,7 +105,7 @@ func (c Spec) InstanceHash() string {
 // an empty raw message.  This is useful for structs where fields are json.RawMessage pointers for bi-directional
 // marshal and unmarshal (value receivers will encode base64 instead of raw json when marshaled), so bi-directional
 // structs should use pointer fields.
-func RawMessage(r *plugin.Any) (raw json.RawMessage) {
+func RawMessage(r *types.Any) (raw json.RawMessage) {
 	if r != nil {
 		raw = json.RawMessage(r.Bytes())
 	}
@@ -112,7 +113,7 @@ func RawMessage(r *plugin.Any) (raw json.RawMessage) {
 }
 
 // RawMessagePtr makes a copy of the Any and make it available as a pointer to a JSON raw message
-func RawMessagePtr(r *plugin.Any) *json.RawMessage {
+func RawMessagePtr(r *types.Any) *json.RawMessage {
 	if r == nil {
 		return nil
 	}
